requirements.txt
gradio>=4.39.0
PyPDF2>=3.0.1
faiss-cpu>=1.12.0
numpy>=2.3.2
google-genai>=1.3.0
python-dotenv>=1.1.0
